requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2
python-dotenv
autopep8
passlib[bcrypt]
python-jose[cryptography]